Stateful CronChecker
